Add CLI help menus to Continuous Integration
This patchs updates the CI testing, adding running the `--help` menu of
each command-line-interface command and command-verb.  (I've personally
found this a helpful catcher in case some entry point is missed by the
unit test coverage.)

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>