Align tested Python versions
This patchs updates the CI testing, adding runs in other current Python
versions.

The versions added---3.8 and 3.9---are added based on the intersection
of the versions Ontospy's dependencies support, and versions not yet at
end-of-life:
https://devguide.python.org/#status-of-python-branches

One notable dependency-based reduction is rdflib no longer supporting
3.6 or older.

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>